test: move SelectBox specs out of src into __tests__
Colocate Jest suites and test helpers under packages/multi-selectbox/__tests__
so production src only contains library implementation.
