+++ packages/multi-selectbox/__tests__/SelectBox.filtering.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, screen } from '@testing-library/react-native'
 import { OPTIONS_MUTABLE } from './test-utils/fixtures'
 import { renderSelectBox } from './test-utils/renderSelectBox'
-import { TEST_IDS } from './testIDs'
+import { TEST_IDS } from '../src/testIDs'
 
 describe('SelectBox filtering', () => {
   it('filters options by input text (case-insensitive)', () => {
