+++ packages/multi-selectbox/__tests__/SelectBox.multi.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, screen } from '@testing-library/react-native'
 import { TouchableOpacity } from 'react-native'
 import { OPTIONS_MUTABLE } from './test-utils/fixtures'
 import { renderSelectBox } from './test-utils/renderSelectBox'
-import { TEST_IDS } from './testIDs'
+import { TEST_IDS } from '../src/testIDs'
 
 describe('SelectBox multi select', () => {
   it('calls onMultiSelect when an option is pressed', () => {
