+++ packages/multi-selectbox/__tests__/test-utils/renderSelectBox.tsx
@@ -1,6 +1,6 @@
 import { render, type RenderOptions } from '@testing-library/react-native'
-import SelectBox from '../SelectBox'
-import type { SelectBoxProps } from '../types'
+import SelectBox from '../../src/SelectBox'
+import type { SelectBoxProps } from '../../src/types'
 
 export function renderSelectBox(props: SelectBoxProps, options?: RenderOptions) {
   return render(<SelectBox {...props} />, options)
